+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Added
+- Support for appearance 'hide-input' in ArcGIS geo widget.
+
 [4.31.4] - 2017-05-03
 ---------------------
 ##### Fixed